Fix budget enforcement bug in EDF-WM
After budget enforcement was removed from
the scheduling plugins and added to Litmus 2010.2,
EDF-WM no longer worked.
The timers were being armed with the execution
cost of the entire task, without considering the
slices.
The bug was fixed by temporarily changing
task_params.exec_cost before arming a timer for a
migratory task, tricking the scheduler.
budget_remaining() compares task_params.exec_cost
with job_params.exec_time (the execution time of the
entire job). So we have to make sure exec_cost is the
cost of all slices until the current one (inclusive).
I had to move slice_budget() to the beginning of the file.