Apply Linux patches to fix 2.6.34 compilation
Current version of GCC finds some errors
in kernel 2.6.34 and doesn't compile it.
I collected some newer kernel patches that fix that.
Don't use this patch with pure Litmus 2010.2,
because it causes a problem in /dev/litmus/log,
making the log unreadable (memory allocation error).
But it works well after we patch feather-trace to
2012.1, as the device allocation procedure is
different.
If you want to use Litmus 2010.2 without feather-trace
2012.1, compile the kernel with an older version of gcc.
The version of Ubuntu 10.0.4 LTS works fine.